backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
aiohttp==3.12.13
beautifulsoup4==4.14.2
blis==1.3.3
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
deep-translator==1.11.4
dnspython==2.7.0
emoji==2.15.0
exceptiongroup==1.3.1
fastapi==0.122.0
geographiclib==2.1
geopy==2.4.1
Jinja2==3.1.6
joblib==1.5.2
kafka-python
langdetect==1.0.9
MarkupSafe==3.0.3
murmurhash==1.0.15
nltk==3.9.2
numpy==1.26.4
openai==2.8.1
openpyxl==3.1.5
packaging==25.0
pandas==2.0.3
preshed==3.0.12
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.4
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
six==1.17.0
shapely
smart_open==7.5.0
soupsieve==2.8
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.49.3
thinc==8.3.9
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
elasticsearch==8.11.0
email-validator
aiokafka
pytest 
pytest-asyncio 
httpx
xradar
fsspec
s3fs
xarray
netCDF4
h5netcdf
matplotlib
pyproj
scipy
